ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.2
pydantic==2.5.0
psycopg2-binary==2.9.9
httpx==0.25.2
python-multipart==0.0.6
jinja2==3.1.2
python-slugify==8.0.1
PyJWT==2.8.0